move routes to root directory and have it so that the use local is part of server config, not just groupme
