+++ bot/src/main/resources/application.conf
@@ -3,19 +3,19 @@ dndbot {
         host = "localhost"
         host = ${?BOT_SERVER_HOST}
 
-        port = 8000
+        port = 7998
         port = ${?PORT}
+
+        use-local = false
+        use-local = ${?GROUPME_BOT_LOCAL}
     }
 
     groupme {
         post-url = "https://api.groupme.com/v3/bots/post"
         post-url = ${?GROUPME_POST_URL}
 
-        bot-id = "bot_id"
+        bot-id = "81f9fa9ce990a921dadb9dcf4e"
         bot-id = ${?GROUPME_BOT_ID}
-
-        use-local = true
-        use-local = ${?GROUPME_BOT_LOCAL}
     }
 
     github {
